requirements.txt
Django==6.0.1
django-ninja==1.5.3
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
email-validator>=2.1.0
python-multipart>=0.0.9
gunicorn>=21.2.0
requests>=2.31.0
openpyxl>=3.1.2